requirements.txt
# Web Framework
fastapi==0.116.2
uvicorn==0.35.0

# Document Processing
python-docx==1.2.0
PyPDF2==3.0.1

# AI/ML Libraries
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15

# Data Processing
pandas==2.3.2
openpyxl==3.1.5

# Data Validation
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1

# Cloud Storage
azure-core==1.35.1
azure-storage-blob==12.26.0

# Utilities
regex==2025.9.1
multi-part